==> 1.data_크롤링.ipynb <==
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
import time

driver = webdriver.Chrome()
driver.get("https://eminwon.geumjeong.go.kr/emwp/gov/mogaha/ntis/web/emwp/cmmpotal/action/EmwpMainMgtAction.do")

selector1 = '#header > div.gnbmiddle > div > div.topmenu > ul > li:nth-child(1) > a'
driver.find_element(By.CSS_SELECTOR, selector1).click()
time.sleep(1)
selector2= '#allmenu > div > div:nth-child(2) > div:nth-child(2) > ul > li:nth-child(2) > a'
driver.find_element(By.CSS_SELECTOR, selector2).click()
time.sleep(1)
selector3= '#pt_deal_state > option:nth-child(4)'
driver.find_element(By.CSS_SELECTOR, selector3).click()
time.sleep(1)
selector4= '#searchBtn'
driver.find_element(By.CSS_SELECTOR, selector4).click()

complain = []
complain_title = []
complain_writer = []
complain_date = []

answer = []
answer_department = []
answer_num = []
answer_writer = []
answer_phone = []
answer_email = []



complain_selector = 'body > main > div > fieldset > table > tbody > tr:nth-child(4) > td'
complain_title_selector = 'body > main > div > fieldset > table > tbody > tr:nth-child(1) > td:nth-child(4)'
complain_writer_selector = 'body > main > div > fieldset > table > tbody > tr:nth-child(2) > td:nth-child(2)'
complain_date_selector = 'body > main > div > fieldset > table > tbody > tr:nth-child(2) > td:nth-child(4)'

answer_selector = 'body > main > div > table > tbody > tr:nth-child(3) > td'
answer_department_selector = 'body > main > div > table > tbody > tr:nth-child(1) > td:nth-child(2)'
answer_num_selector = 'body > main > div > table > tbody > tr:nth-child(1) > td:nth-child(6)'
answer_writer_selector = 'body > main > div > table > tbody > tr:nth-child(2) > td:nth-child(2)'
answer_phone_selector = 'body > main > div > table > tbody > tr:nth-child(2) > td:nth-child(4)'
answer_email_selector = 'body > main > div > table > tbody > tr:nth-child(2) > td:nth-child(6)'


string = ''
crawling = [complain_selector,complain_title_selector,complain_writer_selector,complain_date_selector,answer_selector,answer_department_selector,answer_num_selector,answer_writer_selector,answer_phone_selector,answer_email_selector]
data_list = [complain,complain_title,complain_writer,complain_date,answer,answer_department,answer_num,answer_writer,answer_phone,answer_email]

next_page = '#navigator > a.navi.navi-arrow.navi-arrow-single-right'

su = 0
for k in range(1, 251):
    for i in range(1, 21):
        list_selector = f'#dataSetTb > table > tbody > tr:nth-child({i}) > td.td-list > a'
        driver.find_element(By.CSS_SELECTOR, list_selector).click()
        
        for j in range(10):
            elements = driver.find_elements_by_css_selector(crawling[j])
            
            for element in elements:
                text = element.text
                string += text
                data_list[j].append(string)
            string = ""
        driver.back()
        su += 1
        print(su)
    driver.find_element(By.CSS_SELECTOR, next_page).click()

print(len(answer_department))
print(answer_department[5297])

data_list = [complain,complain_title,complain_writer,complain_date,answer,answer_department,answer_num,answer_writer,answer_phone,answer_email]


df = pd.DataFrame(data_list).transpose()
df = df.dropna()
df

df_column_name = {
    0 : '민원 내용',
    1 : '민원 제목',
    2 : '민원 작성자',
    3 : '민원 작성일',
    4 : '답변 내용',
    5 : '답변 담당부서',
    6 : '답변 접수번호',
    7 : '답변 작성자',
    8 : '답변 전화번호',
    9 : '답변 이메일',
}
df = df.rename(columns = df_column_name)
df

selected_columns = ['민원 내용', '답변 담당부서']
new_df = df[selected_columns]

df.to_csv('output.csv', index = False)

import pandas as pd
df = pd.read_csv("output.csv")

df.info()

import numpy as np
np.unique(df.loc[df['답변 담당부서'].str.endswith('과'), '답변 담당부서'])

display(df.loc[df['답변 담당부서'] == '안전도시국 도시재생과', '민원 내용'])
