==> 3.ai_moding.ipynb <==
import numpy as np
import pandas as pd
from gensim.models.word2vec import Word2Vec
from konlpy.tag import Okt

train = pd.read_csv("output_data.csv", encoding='utf-8')
train

test = pd.read_csv('test.csv')
test = pd.DataFrame(test, columns=['text'])

train['답변 담당부서'].value_counts()

train = train[['민원 내용', '답변 담당부서']]

train.shape

train.info()

from konlpy.tag import Okt

okt = Okt()

# 이랬던 문장이.
print("원본 문장 : ", train.iloc[0]['민원 내용']) 

# 이렇게 바뀝니다. okt.morphs가 형태소 분석기에요.
token_okt = okt.morphs(train.iloc[0]['민원 내용'])
print("형태소 기준으로 tokenize : ", token_okt) 

# 형태소로 토크나이징
temp = []
for i in train['민원 내용']:
    temp.append(okt.morphs(i))
    
temp = pd.Series(temp)
train['token'] = temp

train.info()

from gensim.models import Word2Vec

train

# Word2Vec 모델 학습
wv_model = Word2Vec(train['token'], vector_size = 1000, window = 5, min_count = 5, workers = 4, sg = 0)

import joblib

joblib.dump(wv_model, 'word_model.joblib')

wv_model.wv.vectors.shape

def get_sent_embedding(model, embedding_size, tokenized_words):
    feature_vec = np.zeros((embedding_size,), dtype="float32")
    n_words = 0
    for word in tokenized_words:
        if word in model.wv.key_to_index:
            n_words += 1
            # 임베딩 벡터에 해당 단어의 벡터를 더함
            feature_vec = np.add(feature_vec, model.wv[word])
    # 단어 개수가 0보다 큰 경우 벡터를 단어 개수로 나눠줌 (평균 임베딩 벡터 계산)
    if (n_words > 0):
        feature_vec = np.divide(feature_vec, n_words)
    return feature_vec

# 문장벡터 데이터 셋 만들기
def get_dataset(sentences, model, num_features):
    dataset = list()

    # 각 문장을 벡터화해서 리스트에 저장
    for sent in sentences:
        dataset.append(get_sent_embedding(model, num_features, sent))

    # 리스트를 numpy 배열로 변환하여 반환
    sent_embedding_vectors = np.stack(dataset)
    
    return sent_embedding_vectors

train_data_vecs = get_dataset(train['token'], wv_model, 1000)

from sklearn.model_selection import train_test_split

target = '답변 담당부서'
x = train.drop(target, axis=1)
y = train[target]

x

x = get_dataset(x['token'], wv_model, 1000)

y.value_counts()

x_train, x_val, y_train, y_val = train_test_split(x, y, test_size=0.15)

from sklearn.ensemble import RandomForestClassifier
model = RandomForestClassifier()

model.fit(x_train, y_train)

y_pred = model.predict(x_val)

from sklearn.metrics import accuracy_score
from sklearn.metrics import f1_score

accuracy_score(y_val, y_pred)

f1_score(y_val, y_pred, average="macro")

import joblib
joblib.dump(model, 'model.joblib')

import pandas as pd

data = pd.read_csv("output_data.csv")

data.head()

data['답변 담당부서'].value_counts()

data.loc[data['답변 담당부서'] == '안전도시국 건설과','민원 내용'].iloc[0]

data.loc[data['답변 담당부서'] == '경제환경국 공원녹지과','민원 내용'].iloc[25]
